+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - Add support for kafka-node >= 4.0.0.
+- Bunyan: support for capturing the error message from a nested error object (attribute `err`) given in the fields argument.
 
 ## 1.70.0
 - Enable uncaught exception reporting in Node 12.x.